FIX: checkstyle and incompatible datatypes

+++ src/main/java/de/hhu/propra/sharingplatform/faker/ItemFaker.java
@@ -18,8 +18,8 @@ public class ItemFaker {
         Item item = new Item();
         item.setName(faker.space().nasaSpaceCraft());
         item.setDescription(faker.lorem().paragraph(1));
-        item.setBail(faker.number().numberBetween(20, 999));
-        item.setPrice(faker.number().numberBetween(5, 50));
+        item.setBail((double)faker.number().numberBetween(20, 999));
+        item.setPrice((double)faker.number().numberBetween(5, 50));
         item.setAvailable(true);
         item.setLocation(faker.address().cityName());
         item.setDeleted(false);
